requirements.txt
# Python依赖包 - 优化版本
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3  # 升级到支持 Flask 3.0.0 的版本
Flask-WTF>=1.2.0
Flask-Mail>=0.9.1
Werkzeug>=2.3.0,<3.1.0
SQLAlchemy>=2.0.23
PyMySQL>=1.1.0
python-dotenv>=1.0.0
Flask-Migrate>=4.0.5
Flask-JWT-Extended>=4.5.0
marshmallow>=3.20.1
cryptography>=41.0.3
pytest>=8.0.0
pytest-flask>=1.3.0
bcrypt>=4.1.1
pytest-mock>=3.14.0